requirements.txt
google-api-python-client==2.143.0
youtube_transcript_api==0.6.2
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.5
pydantic==2.8.2
opencv-python==4.10.0.84
scikit-learn==1.5.1
Pillow==10.4.0
open_clip_torch==2.26.1
torch>=2.3.0
yt-dlp==2024.08.06
streamlit==1.37.1
textblob==0.18.0.post0
langcodes==3.4.0
PyYAML==6.0.2
